Moving map drawer to utils
