+++ pathgenerator/runner.py
@@ -3,7 +3,7 @@ from pathgenerator.utils.data_fetcher import DataFetcher
 from PIL import Image, ImageDraw
 from threading import Thread
 
-import pathgenerator.map_drawer as map_drawer
+import pathgenerator.utils.map_drawer as map_drawer
 import pathgenerator.utils.imgur_uploader as imgur_uploader
 from pathgenerator.config import ALL_WORLDS, DB_HOST, DB_DATABASE, DB_USER, DB_PASSWORD
